+++ src/main/java/localgaji/albbaim/__core__/auth/AuthUserArgumentResolver.java
@@ -2,8 +2,8 @@ package localgaji.albbaim.__core__.auth;
 
 import localgaji.albbaim.__core__.exception.CustomException;
 import localgaji.albbaim.__core__.exception.ErrorType;
-import localgaji.albbaim.auth.user.User;
-import localgaji.albbaim.auth.user.UserRepository;
+import localgaji.albbaim.user.User;
+import localgaji.albbaim.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
