+++ src/main/java/localgaji/albbaim/user/UserController.java
@@ -1,10 +1,10 @@
-package localgaji.albbaim.auth.user;
+package localgaji.albbaim.user;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import localgaji.albbaim.__core__.ApiUtil;
 import localgaji.albbaim.__core__.auth.AuthUser;
-import localgaji.albbaim.auth.user.userDTO.ResponseUser.GetMyInfoResponse;
+import localgaji.albbaim.user.userDTO.ResponseUser.GetMyInfoResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
