+++ src/main/java/localgaji/albbaim/user/userDTO/ResponseUser.java
@@ -1,8 +1,8 @@
-package localgaji.albbaim.auth.user.userDTO;
+package localgaji.albbaim.user.userDTO;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import localgaji.albbaim.auth.user.User;
-import localgaji.albbaim.auth.user.UserType;
+import localgaji.albbaim.user.User;
+import localgaji.albbaim.user.UserType;
 import lombok.Getter;
 
 public class ResponseUser {
