+++ src/main/java/localgaji/albbaim/workplace/WorkplaceService.java
@@ -4,8 +4,8 @@ import localgaji.albbaim.__core__.exception.CustomException;
 import localgaji.albbaim.__core__.exception.ErrorType;
 import localgaji.albbaim.workplace.invitation.Invitation;
 import localgaji.albbaim.workplace.invitation.InvitationService;
-import localgaji.albbaim.auth.user.User;
-import localgaji.albbaim.auth.user.UserService;
+import localgaji.albbaim.user.User;
+import localgaji.albbaim.user.UserService;
 import localgaji.albbaim.workplace.workplaceDTO.ResponseWorkplace.GetMyWorkplaceResponse.UserListDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
