+++ src/test/java/localgaji/albbaim/auth/AuthServiceTest.java
@@ -6,8 +6,8 @@ import localgaji.albbaim.auth.authDTO.RequestAuth.LoginRequest;
 import localgaji.albbaim.auth.authDTO.RequestAuth.SignUpRequest;
 import localgaji.albbaim.auth.oauth.kakaoAuth.KakaoAuth;
 import localgaji.albbaim.auth.oauth.kakaoAuth.KakaoAuthService;
-import localgaji.albbaim.auth.user.User;
-import localgaji.albbaim.auth.user.UserService;
+import localgaji.albbaim.user.User;
+import localgaji.albbaim.user.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
